server/requirements.txt
Flask 
flask-cors 
flask-bcrypt 
flask-jwt-extended 
pymongo 
python-dotenv
langchain 
langchain-groq